Add files via upload
Please check the arc_calculations.py, there is a function for calculating the CC of two segments. I didn't find a way to read the segment markers from the code movive_client.py cos I cannot run it yet. 

Please be reminded that the input of the function calculate_curvature requires the segments_points array in the format of [end1, centre, end2], where the end1 should be the marker closest to the rigid body.

To output CC has positive/negative, which the concave is negative.